dreco/commands2.py
import subprocess
import requests
import dns.resolver
import socket

def get_ip_from_domain(domain):
    """Fetches the IP address associated with the given domain."""
    try:
        ip_address = socket.gethostbyname(domain)
        return ip_address
    except socket.gaierror:
        print(f"[ERROR] Could not resolve domain: {domain}")
        return None

def get_a_records(domain):
    # Implementation of A records retrieval
    pass

def get_ns_records(domain):
    # Implementation of NS records retrieval
    pass

def get_aaaa_records(domain):
    # Implementation of AAAA records retrieval
    pass

def get_mx_records(domain):
    # Implementation of MX records retrieval
    pass

def sublist3r_scan(domain):
    # Implementation of Sublist3r scanning
    pass

def nmap_scan(target_ip):
    """Runs Nmap on the specified IP address."""
    print(f"Running Nmap scan on {target_ip}...")
    command = ['nmap', '-sV', target_ip]  # Example options, adjust as needed
    result = subprocess.run(command, capture_output=True, text=True)
    print(result.stdout)

def fetch_robots_txt(domain):
    # Implementation of robots.txt fetching
    pass

def run_dnsrecon(domain):
    """Runs DNSRecon on the specified domain."""
    print(f"Running DNSRecon for {domain}...")
    command = ['dnsrecon', '-d', domain, '-a']  # Adjust options as needed
    result = subprocess.run(command, capture_output=True, text=True)
    print(result.stdout)

def run_dnsenum(domain):
    """Runs DNSEnum on the specified domain."""
    print(f"Running DNSEnum for {domain}...")
    command = ['dnsenum', domain]  # Adjust options as needed
    result = subprocess.run(command, capture_output=True, text=True)
    print(result.stdout)

