requirements.txt
requests
dnspython
sublist3r
dnsrecon
python-nmap
